Implement Veilederinfo-component
The previous version of the app used an outdated component (DialogBox) to show
information messages in certain sections of the application. This commit replaces
this component and its usage with the new Veilederinfo-component.

Also made a change to LinkWithIcon so that it now uses the internal CustomSVG-component
instead of depending on nav-frontend-ikoner-assets to display icons.