Update nav-frontend-skjema
Updated nav-frontend-skjema and nav-frontend-skjema-style versions
as it has a bugfix for BekreftCheckboxPanel toggling on and off value
when user clicks enter while it has focus, which in our case
resulted in set value to true, followed by submitting the form. This
no longer happens as this custom event handler on Enter-click has
been removed from the newest version of nav-frontend-skjema.

Also fixes an alignment issue with label and checkbox of
BekreftCheckboxPanel in Firefox and Safari.