Better UX on Step3
This fix makes sure utenlandsopphold added to lists for tidligerePerioder and senerePerioder
are not reset when changing values of form-elements elsewhere in the form, including the RadioPanels
related to the lists themselves.

This commit also makes sure that we do not send tidligerePerioder- and senerePerioder-data from the
state of the application to foreldrepengesoknad-api when the user has added an Utenlandsopphold to any of the lists, followed by
stating they have been in norway / will live in norway the previous and/or next 12 months.