Fødselsdato-field fix
User should not be able to add fødselsdato more than 3 years back in time.
This commit makes sure to disable dates in datepicker that are more than 3 years
in the past. Also makes sure to present user with validation-error if attempting
to add date too far in the past via the input text field.