Update is-valid-fodselsnummer pkg
Previous version of is-valid-fodselsnummer package was using fat arrow syntax, resulting
in application breaking on IE11 and below. This was fixed in the newest version, allowing us
to remove a previous modification to babel-loader to make that specific dependency be included
in the transpilation.

Also, it seems like package-lock has not been updated recently, which has been done in this
commit aswell.